Checkpoints/LightRefill
Added Checkpoint and LightRefill prefabs. Functionality should be built
into the player controller (there will be debug messages, don’t worry
about that). Checkpoint resets when you exit the level, but not
restart, so when testing you’ll respond at the checkpoint. Refill
immediately refills lamp to max. So far, no sprites, so the objects are
invisible, but the prefabs are in there.